Reorganize Solidity abstraction implementations
